Move framework-specific preprocessing from classification to estimator level
